+++ CHANGELOG.md
@@ -5,7 +5,12 @@
 ### Added
 
 - Added a `--check` option to verify if dependencies are already sorted.
-- `POETRY_SORT_ENABLED` environment variable to enable or disable sorting in `poetry init`, `poetry add` commands.
+- New environment variable `POETRY_SORT_ENABLED` to enable or disable sorting in `poetry init`, `poetry add` commands.
+- New environment variable `POETRY_SORT_MOVE_OPTIONALS_TO_BOTTOM` to move optional packages to the bottom.
+
+### Removed
+
+- Support for `poetry remove` command.
 
 ## [0.1.1] - 2022-10-25